use the correct core folder
