+++ server/src/packages/submission/submission.repository.ts
@@ -1,5 +1,5 @@
 import { Option } from 'funfix';
-import { uuidType } from '../../core';
+import { uuidType } from '../core';
 
 export class SubmissionId extends uuidType<'SubmissionId'>() {}
